Refactor rules multi select into its own component
