+++ src/components/Input.tsx
@@ -1,6 +1,8 @@
 import { twMerge } from "tailwind-merge";
 import { Game, Rule } from "~utils/hooks/rules";
 import { useCallback } from "react";
+import { IconButton } from "./Button";
+import { TrashIcon } from "@heroicons/react/24/outline";
 
 export type InputBaseProps = React.HTMLProps<HTMLInputElement>;
 
@@ -107,5 +109,84 @@ export const RulesSelect: React.FC<RulesSelectProps> = ({
         </optgroup>
       ))}
     </Select>
+  );
+};
+
+export type RulesMultiSelectProps = {
+  game: Game;
+  value: Rule[];
+  onChange: (rules: Rule[]) => void;
+};
+
+export const RulesMultiSelect: React.FC<RulesMultiSelectProps> = ({
+  game,
+  value,
+  onChange,
+}) => {
+  const onPickOtherRule = useCallback(
+    (e: React.ChangeEvent<HTMLSelectElement>) => {
+      const group = e.target.selectedOptions[0].dataset.rulegroup;
+      if (!group) return;
+
+      const rule = game?.ruleGroups
+        .find((g) => g.name === group)
+        ?.rules.find((r) => r.rule === e.target.value);
+
+      if (!rule) return;
+      if (value.some((r) => r.rule === rule.rule)) return;
+      onChange([...value, rule]);
+    },
+    [game, value]
+  );
+
+  const onRemoveRule = useCallback(
+    (rule: Rule) => {
+      const rules = value.filter((r) => r.rule !== rule.rule);
+      onChange(rules);
+    },
+    [value]
+  );
+
+  return (
+    <>
+      <Select
+        className="max-w-full w-full mt-2"
+        value={""}
+        onChange={onPickOtherRule}
+      >
+        <option>Pick Rule</option>
+        {game?.ruleGroups.map((group) => (
+          <optgroup label={group.name} key={group.name}>
+            {group.rules.map((rule) => (
+              <option
+                value={rule.rule}
+                data-rulegroup={group.name}
+                key={rule.rule}
+              >
+                {rule.rule} {rule.description}
+              </option>
+            ))}
+          </optgroup>
+        ))}
+      </Select>
+      <ul className="mt-4 flex flex-wrap gap-2">
+        {value.map((rule) => (
+          <li
+            key={rule.rule}
+            className="p-2 flex w-full items-center bg-zinc-800 rounded-md"
+          >
+            <p className="flex-1 mr-1">
+              <strong className="font-mono mr-2">{rule.rule}</strong>
+              <span>{rule.description}</span>
+            </p>
+            <IconButton
+              className="bg-transparent"
+              icon={<TrashIcon height={24} />}
+              onClick={() => onRemoveRule(rule)}
+            ></IconButton>
+          </li>
+        ))}
+      </ul>
+    </>
   );
 };
